first full draft of extent metadata caching

+++ server/src/unifyfs_group_rpc.c
@@ -55,7 +55,7 @@ static int forward_child_request(void* input_ptr,
     int ret = UNIFYFS_SUCCESS;
 
     /* call rpc function */
-    double timeout_ms = margo_service_timeout_msec;
+    double timeout_ms = (double) margo_service_timeout_msec;
     hg_return_t hret = margo_iforward_timed(chdl, input_ptr, timeout_ms, creq);
     if (hret != HG_SUCCESS) {
         LOGERR("failed to forward request(%p) - %s", creq,
